Update 20shares - correct s
For Shared folder group privileges to be correctly converted to afp.conf entries, the @ must be part of the group entry within the hyphens and not in front of it (WRONG: @"groupname", CORRECT: "@groupname", see http://netatalk.sourceforge.net/3.0/htmldocs/afp.conf.5.html). Else afp will not present to clients.